Refactor UserResolveExecutor to use MessageDTO and update error handling
