+++ components/org.wso2.carbon.identity.application.authentication.handler.identifier/src/main/java/org/wso2/carbon/identity/application/authentication/handler/identifier/UserResolveExecutor.java
@@ -30,7 +30,7 @@ import org.wso2.carbon.identity.flow.execution.engine.graph.Executor;
 import org.wso2.carbon.identity.flow.execution.engine.model.ExecutorResponse;
 import org.wso2.carbon.identity.flow.execution.engine.model.FlowExecutionContext;
 import org.wso2.carbon.identity.flow.mgt.model.ExecutorDTO;
-import org.wso2.carbon.identity.flow.mgt.model.Message.MessageType;
+import org.wso2.carbon.identity.flow.mgt.model.MessageDTO.MessageType;
 import org.wso2.carbon.identity.flow.mgt.model.NodeConfig;
 import org.wso2.carbon.identity.multi.attribute.login.mgt.ResolvedUserResult;
 import org.wso2.carbon.user.api.UserStoreException;
@@ -152,14 +152,12 @@ public class UserResolveExecutor implements Executor {
                         IdentifierHandlerConstants.AccountLockedReason lockedReason =
                                 IdentifierHandlerConstants.AccountLockedReason.fromReason(reason);
                         executorResponse.setResult(STATUS_RETRY);
-                        executorResponse.setErrorMessage(lockedReason.getMessage());
                         executorResponse.addMessage(MessageType.ERROR, lockedReason.getMessage(),
                                 lockedReason.getI18nKey());
                         return executorResponse;
                     }
                     if (Boolean.parseBoolean((String) claimMap.get(FrameworkConstants.ACCOUNT_DISABLED_CLAIM_URI))) {
                         executorResponse.setResult(STATUS_RETRY);
-                        executorResponse.setErrorMessage(IdentifierHandlerConstants.ACCOUNT_DISABLED);
                         executorResponse.addMessage(MessageType.ERROR, IdentifierHandlerConstants.ACCOUNT_DISABLED,
                                 IdentifierHandlerConstants.ACCOUNT_DISABLED_I18N_KEY);
                         return executorResponse;
@@ -176,7 +174,6 @@ public class UserResolveExecutor implements Executor {
                 }
                 if (isNotifyUserExistenceEnabled(context)) {
                     executorResponse.setResult(STATUS_RETRY);
-                    executorResponse.setErrorMessage(IdentifierHandlerConstants.INVALID_IDENTIFIER);
                     executorResponse.addMessage(MessageType.ERROR, IdentifierHandlerConstants.INVALID_IDENTIFIER,
                             IdentifierHandlerConstants.INVALID_IDENTIFIER_I18N_KEY);
                     return executorResponse;
